+++ src/components/domain/my-quotes/QuotesLoading.tsx
