+++ src/lib/api/estimate/getClientPendingQuote.ts
@@ -1,0 +1,9 @@
+import { tokenFetch } from "@/lib/utils";
+
+export async function fetchClientPendingQuotes() {
+   try {
+      return await tokenFetch("/estimates/pending", { method: "GET" });
+   } catch (e) {
+      throw e;
+   }
+}
