+++ src/lib/api/estimate/getClientReceivedQuote.ts
@@ -1,0 +1,11 @@
+import { tokenFetch } from "@/lib/utils";
+
+export async function fetchClientReceivedQuotes(category: string) {
+   try {
+      return await tokenFetch(`/estimates/received?category=${category}`, {
+         method: "GET",
+      });
+   } catch (e) {
+      throw e;
+   }
+}
