+++ src/lib/types/index.ts
@@ -3,3 +3,4 @@ export * from "./profile.types";
 export * from "./review.types";
 export * from "./request.types";
 export * from "./mover.types";
+export * from "./estimate.types";
